requirements.txt
diffusers~=0.17.1
transformers~=4.30.2
dataclasses~=0.6
Pillow~=10.0.0
PyYAML~=6.0
torch~=2.0.1
moviepy~=1.0.3
opencv-python~=4.8.0.74
accelerate~=0.20.3
suno-bark @ git+https://github.com/suno-ai/bark.git
pydub~=0.25.1
